legacy/containers/reduce/requirements.txt
# Reduce Container Dependencies
# These are the runtime dependencies for the reduce phase container

# Core ML/NLP
bertopic>=0.16.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
hdbscan>=0.8.33
scikit-learn>=1.3.0
numpy>=1.24.0
spacy>=3.7.0

# Data Processing
pandas>=2.0.0
pyyaml>=6.0

# AWS SDK
boto3>=1.28.0

# Utilities
tqdm>=4.66.0
